+++ apps/fluent-tester/src/TestComponents/Menu/consts.ts
@@ -14,3 +14,5 @@ export const MENUITEM_TEST_LABEL = 'Test MenuItem2 - No Accessibility Label'; //
 
 export const MENU_ON_OPEN = 'Menu_On_Open'; // For testing the open functionality of Menu
 export const MENU_ON_CLOSE = 'Menu_On_Close'; // For testing the open functionality of Menu
+
+export const MENU_DEFOCUS_BUTTON = 'Menu_Defocus_Button'; // Used to move focus away from menu button