+++ packages/components/Badge/src/CounterBadge/CounterBadgeTokens.ts
@@ -4,6 +4,7 @@ import { CounterBadgeTokens } from './CounterBadge.types';
 
 export const counterBadgeTokens: TokenSettings<CounterBadgeTokens> = () =>
   ({
+    shadowToken: undefined,
     dot: {
       width: 6,
       minHeight: 6,