+++ packages/utils/adapters/src/adapters.win32.ts
@@ -157,8 +157,6 @@ const _imageMask: IFilterMask<IImageProps> = {
   defaultSource: true,
   blurRadius: true,
   capInsets: true,
-  width: true,
-  height: true,
   fadeDuration: true,
   nativeID: true,
 };