+++ packages/utils/adapters/src/adapters.windows.ts
@@ -70,6 +70,11 @@ const _viewMask: IFilterMask<IViewProps> = {
   tvParallaxShiftDistanceY: true,
   tvParallaxTiltAngle: true,
   focusable: true,
+  tooltip: true,
+  tabIndex: true,
+  enableFocusRing: true,
+  onMouseEnter: true,
+  onMouseLeave: true,
 };
 
 const _textMask: IFilterMask<ITextProps> = {
@@ -85,6 +90,7 @@ const _textMask: IFilterMask<ITextProps> = {
   accessibilityValue: true,
   accessibilityViewIsModal: true,
   accessible: true,
+  disabled: true,
   onAccessibilityAction: true,
   onAccessibilityEscape: true,
   onAccessibilityTap: true,
@@ -154,8 +160,6 @@ const _imageMask: IFilterMask<IImageProps> = {
   defaultSource: true,
   blurRadius: true,
   capInsets: true,
-  width: true,
-  height: true,
   fadeDuration: true,
   nativeID: true,
 };